fix: rename footer & header folders according to codestyle
